requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv